+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Version 1.16.2:
+
+- `A > B` adds a layout constraint, but not a real connection.
+
 ## Version 1.16.1.post1:
 
 - Fix: When using attribs alias for stores or channel, unquote attribs